AMP-7373 merged from 1.15
git-svn-id: file:///svn/branches/AMP_1_16_RELEASE/amp@15373 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/util/LuceneUtil.java
@@ -300,11 +300,13 @@ public class LuceneUtil implements Serializable {
     	}
     	
     	if (idxDir.exists()){
-			logger.info("Found, deleting ...");
-			if (deleteDirectory(idxDir))
-				logger.info("Done");
-			else
-				logger.info("Can't delete!");
+    		if (deleteIndex){
+    			logger.info("Found, deleting ...");
+    			if (deleteDirectory(idxDir))
+    				logger.info("Done");
+    			else
+    				logger.info("Can't delete!");
+    		}
 		}
 		else{
 			logger.info("Not found ... will be generated!");
